fix: return to correct view when canceling AC feedback input
When pressing 'f' on an AC in iteration detail view and then ESC,
the code was hardcoded to return to ViewACList which requires
m.currentTrack to be set. Since iteration view doesn't set
m.currentTrack, this caused a nil pointer dereference.

Fixed by using m.previousViewMode instead of hardcoding ViewACList,
so the UI returns to whichever view the user came from. Also fixed
the Enter (submit) path to reload the appropriate data based on
the previous view context.